Update exception handling for wannacry_tlscan.py
After doing some internal scans on SMBv1 (but patched) systems, it seems that the script failed to determine the state of the system. After looking into the script, it seemed that the "except: pass" simply skipped the exception without handling it. Printing the exception made things easier. I wrote an exception class for you (for future stuff as well) which can handle the missing piece. Now, in-cases where the error is being passed as exception, the script can still identify that the system is not vulnerable and prompt to the user.

I encorage you to make the script more "debuggy" for others to follow.

Ido